Updated
Rewrote the dynamic FPS mod and fixed most of the known issues.
Created a function to calculate all in-game timings.

Created a function to handle gameplay between 30 and 45 FPS.

Created a function to handle gameplay between 46 and 60 FPS.

Created a function to detect when the player climbs a ladder.

Created a function to detect when the player performs an attack.

Smoothed all FPS values to prevent stuttering during rolls.